feat(shared): organize shared type definitions and add missing types
Split shared types into separate files (types/repository.ts, types/api.ts)
for better organization. Add MetricsDaily, Language, SortBy, PaginationMeta
types from Issue #62 spec. All existing imports remain backward compatible
via re-exports from index.ts.

Closes #62
